requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy>=2.0
alembic>=1.12
psycopg2-binary
authlib
google-auth
httpx
pyjwt
starlette
python-multipart
itsdangerous